# openapi.yaml
openapi: 3.1.0
info:
  title: IFRS9 API
  version: 1.0.0
paths:
...
  /request-access:
    post:
      tags:
      - auth
      summary: Request Access
      description: Request user access by submitting email for verification
      operationId: request_access_request_access_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /submit-admin-request:
    post:
      tags:
      - auth
      summary: Submit Admin Request
      description: Submit admin email for verified access request
      operationId: submit_admin_request_submit_admin_request_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessRequestSubmit'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/requests:
    get:
      tags:
      - admin
      summary: Get Access Requests
      description: Get all access requests
      operationId: list_all_access_requests
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AccessRequestResponse'
                type: array
                title: Response List All Access Requests
      security:
      - OAuth2PasswordBearer: []
  /login:
    post:
      tags:
      - auth
      summary: Login
      description: Login using email and password
      operationId: login_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /verify-email/{token}:
    get:
      tags:
      - auth
      summary: Verify Email
      description: Enter email verification token to verify email address
      operationId: verify_email_verify_email__token__get
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Token not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/requests/{request_id}:
    put:
      tags:
      - admin
      summary: Update Access Request
      description: Update access requests
      operationId: update_specific_access_request
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessRequestUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Request not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - admin
      summary: Delete Access Request
      description: Delete access requests
      operationId: delete_access_request_admin_requests__request_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      responses:
        '204':
          description: Successful Response
        '404':
          description: Request not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /set-password/{token}:
    post:
      tags:
      - auth
      summary: Set Password
      description: Set password using invitation token
      operationId: set_password_set_password__token__post
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          title: Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordSetup'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Token not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AccessRequestResponse:
      properties:
        id:
          type: integer
          title: Id
        email:
          type: string
          format: email
          title: Email
        admin_email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Admin Email
        status:
          type: string
          title: Status
        created_at:
          type: string
          format: date-time
          title: Created At
        is_email_verified:
          type: boolean
          title: Is Email Verified
      type: object
      required:
      - id
      - email
      - status
      - created_at
      - is_email_verified
      title: AccessRequestResponse
    AccessRequestSubmit:
      properties:
        email:
          type: string
          format: email
          title: Email
        admin_email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Admin Email
      type: object
      required:
      - email
      title: AccessRequestSubmit
    AccessRequestUpdate:
      properties:
        status:
          $ref: '#/components/schemas/RequestStatus'
        role:
          anyOf:
          - $ref: '#/components/schemas/UserRole'
          - type: 'null'
      type: object
      required:
      - status
      title: AccessRequestUpdate
...
    EmailVerificationRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
      type: object
      required:
      - email
      title: EmailVerificationRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
    LoginResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          pattern: ^bearer$
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          exclusiveMinimum: 0.0
          title: Expires In
        user:
          $ref: '#/components/schemas/UserModel'
      type: object
      required:
      - access_token
      - expires_in
      - user
      title: LoginResponse
...
    PasswordSetup:
      properties:
        password:
          type: string
          minLength: 8
          title: Password
          example: MyS3cur3Pwd
        confirm_password:
          type: string
          title: Confirm Password
      type: object
      required:
      - password
      - confirm_password
      title: PasswordSetup
...
    RequestStatus:
      type: string
      enum:
      - pending
      - approved
      - denied
      - flagged
      title: RequestStatus
...
    UserModel:
      properties:
        id:
          type: integer
          title: Id
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        email:
          type: string
          format: email
          title: Email
        recovery_email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Recovery Email
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
          title: Is Active
        access_request_status:
          anyOf:
          - $ref: '#/components/schemas/RequestStatus'
          - type: 'null'
      type: object
      required:
      - id
      - email
      - role
      - is_active
      title: UserModel
...
    UserRole:
      type: string
      enum:
      - admin
      - analyst
      - reviewer
      - user
      title: UserRole
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: token